requirements.txt
matplotlib>=3.3.3
numpy>=1.19.5
pyserial>=3.5
python-dateutil>=2.8.1
